pyproject.toml
[tool.poetry]
name = "image_to_image"
version = "0.1.0"
description = ""
authors = ["moarshy <marshath@gmail.com>"]
readme = "README.md"
packages = [
    { include = "image_to_image", from = "." }
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pillow = "^10.2.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.2"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
